+++ TameableCreatures/Properties/AssemblyInfo.cs
@@ -1,9 +1,10 @@
-using System.Reflection;
+﻿using System.Reflection;
 
 [assembly: AssemblyCompany("TameableCreatures")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("0.7.0.0")]
-[assembly: AssemblyInformationalVersion("0.7.0")]
+[assembly: AssemblyFileVersion("0.7.1.0")]
+[assembly: AssemblyInformationalVersion("0.7.1")]
 [assembly: AssemblyProduct("TameableCreatures")]
 [assembly: AssemblyTitle("TameableCreatures")]
-[assembly: AssemblyVersion("0.7.0.0")]
+[assembly: AssemblyVersion("0.7.1.0")]
+